the_correction_run_started

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: faventur <faventur@student.42mulhouse.fr>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/23 18:26:26 by faventur          #+#    #+#             */
-/*   Updated: 2022/02/25 11:17:13 by faventur         ###   ########.fr       */
+/*   Updated: 2022/02/26 11:23:13 by faventur         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!n)
